+++ moira_lists/admin.py
@@ -5,6 +5,8 @@ from moira_lists.models import MoiraList
 
 
 class MoiraListAdmin(admin.ModelAdmin):
+    """Admin for Moira Lists"""
+
     model = MoiraList
     search_fields = ("name",)
