+++ moira_lists/tasks.py
@@ -3,7 +3,7 @@ import logging
 from django.contrib.auth import get_user_model
 
 from moira_lists.models import MoiraList
-from moira_lists.moira_api import update_user_moira_lists, update_moira_list_users
+from moira_lists import moira_api
 from open_discussions.celery import app
 
 User = get_user_model()
@@ -18,7 +18,7 @@ def update_user_moira_lists(user_id):
     Args:
         user_id (int): User id
     """
-    update_user_moira_lists(User.objects.get(id=user_id))
+    moira_api.update_user_moira_lists(User.objects.get(id=user_id))
 
 
 @app.task
@@ -30,4 +30,4 @@ def update_moira_list(name):
         name (str): Moira list name
     """
     moira_list, _ = MoiraList.objects.get_or_create(name=name)
-    update_moira_list_users(moira_list)
+    moira_api.update_moira_list_users(moira_list)
